Limit parrallel image decoding to available physical RAM

+++ src/core/vpinball.cpp
@@ -2407,8 +2407,8 @@ int CALLBACK MyCompProcMemValues(LPARAM lSortParam1, LPARAM lSortParam2, LPARAM 
    const SORTDATA * const lpsd = (SORTDATA *)lSortOption;
    const Texture * const t1 = (Texture *)lSortParam1;
    const Texture * const t2 = (Texture *)lSortParam2;
-   const unsigned int t1_size = t1->GetEstimatedGPUSize();
-   const unsigned int t2_size = t2->GetEstimatedGPUSize();
+   const size_t t1_size = t1->GetEstimatedGPUSize();
+   const size_t t2_size = t2->GetEstimatedGPUSize();
    if (lpsd->sortUpDown == 1)
       return (int)(t1_size - t2_size);
    else
